+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - The "X of Y corpora selected" phrase is not properly translated [#408](https://github.com/spraakbanken/korp-frontend/issues/408)
 - Empty localization strings sometimes render as localization key [#410](https://github.com/spraakbanken/korp-frontend/issues/410)
 - Wider filter lists [#412](https://github.com/spraakbanken/korp-frontend/issues/412)
+- Alphabetic sorting of statistics columns [#37](https://github.com/spraakbanken/korp-frontend/issues/37)
 
 ## [9.7.1] - 2024-09-18